update build history when build finishes
previously, viewing an in-flight build, waiting for it to finish, and
then switching away would revert its history item to its pre-finished
state.

lots of this was refactoring to decouple the Build module from the
BuildOutput module. the Build model used to defer to the BuildOutput
model for build info, as the BuildOutput would update its own as events
were handled. but this made the lifecycle confusing, as BuildOutput
really had no need to update its model, and it meant when the
BuildOutput goes away, all previously gained info (e.g. the build's
status) went away with it.

so, BuildOutput.update now returns a third parameter, indicating
messages to send to the parent component. one of these messages is
BuildStatus, which the Build module now handles by updating its own
model *and* the history entry.

other refactors along the way:

* Action -> Msg
* explicit exposing (..) lists
* better encapsulate subscriptions and messages