implement keyboard shortcut 'f' for fitting pipeline to canvas
this is really only possible by using javascript to interact
with the d3 node. so I created a function in index.js called
`resetPipelineFocus` which removes the `transform` attribute
of the d3 node. This effectively refocuses the whole thing and
fits it back to the canvas.

next step was to add this as a port in index.html, and then add that
port to the SubPage and the Pipeline models.

then I added another keyboard press subscription which uses a new msg
KeyPressed, which runs that port resetFocus function when the key
pressed is 'f' or 'F'

concourse/concourse#1389