Empty search now changes URL to /? instead of /
Pushing an empty URL causes the history to not get modified. If we
really wanted to clean the URL of any trace of the search, we could
apparently do so by pushing location.pathname. I'm not sure this is
worth the additional work, so let's just render an empty QueryString.