fix blinking durations in dashboard
no need to use a Maybe Time -- the only way to not know what time it is
is before the very first call to ask the browser. At this point you wont
even have any API data so there will be nothing to put durations on!

thus it is safe to couple the API data and the clock time together.

concourse/concourse#2580

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>